[Fix] Use Insets for Expanded SubappApp Content
While the expanded subapp content offset calculation for grid alignment
worked it broke the connection clipping. The result was that connections
where not ending at the interface bar. With this change now the offset
is applied as insets to expanded SubApp content figure, which fixes the
connection clipping and provides the right offset.
